usage for bjam
Hello, I can't find another way to raise an issue / ask questions. Thanks for adding support for this. Do you have any docs for it? Also are you able to provide guidance on building and using it at all - do you expect it to work with either premake4 or 5 with a standard jam setup?

thanks - Nick